+++ cruiser-game/core/src/com/masonorovic/cruiser/objects/car/engine/Supercharger.java
@@ -2,7 +2,7 @@ package com.masonorovic.cruiser.objects.car.engine;
 
 public class Supercharger extends ForcedInductionSystem {
     public Supercharger(String codeName,
-                        float maxBoostPressure,
+                        int maxBoostPressure,
                         float[] torqueCurve,
                         float damage) {
         super(codeName, maxBoostPressure, torqueCurve, damage);
