+++ cruiser-game/core/src/com/masonorovic/cruiser/objects/car/engine/Turbocharger.java
@@ -4,7 +4,7 @@ public class Turbocharger extends ForcedInductionSystem {
     private Intercooler intercooler;
 
     public Turbocharger(String codeName,
-                        float maxBoostPressure,
+                        int maxBoostPressure,
                         float[] torqueCurve,
                         float damage,
                         Intercooler intercooler) {
